[CP-75] feat (Identity): enhance debug mode logic

+++ Identity.Service/src/main.cpp
@@ -1,7 +1,6 @@
 #if !defined(NDEBUG)
 #include "ApiUtils.h"
 #endif
-
 #include "ContactTable.h"
 #include "Filters.h"
 #include "Handlers.h"
@@ -50,6 +49,12 @@ static void setupEndpoints()
         .registerHandler("/api/user/searchUsers?requestorId={requestorId}&searchedUsername={searchedUsername}",
                          &findUsersWithContactCheck,
                          {Get});
+        // TODO: PATCH: update user status
+        // TODO: PATCH: update user data
+        // TODO: PATCH: update user password
+        // TODO: PATCH: update user avatar
+        // TODO: DELETE: remove user avatar
+        // TODO: DELETE: delete user
 }
 
 static void serviceSetup()
@@ -82,30 +87,79 @@ static void setCorsPolicy(const HttpRequestPtr &req, const HttpResponsePtr &resp
 #if !defined(NDEBUG)
 static void addTestData()
 {
-    LOG_INFO << "Debug mode enabled. Clearing tables. Adding test data.";
+    LOG_INFO << "Debug mode enabled. Using test data.";
 
-    auto dbClient = app().getDbClient();
-    auto futureContact = dbClient->execSqlAsyncFuture("TRUNCATE TABLE \"Contact\" CASCADE;");
-    auto futureUser= dbClient->execSqlAsyncFuture("TRUNCATE TABLE \"User\" CASCADE;");
+    const std::string rootName = "root_user";
+    LOG_WARN << "DO NOT DELETE ROOT USER \"" << rootName << "\". OTHERWISE TABLES WILL BE RECREATED.";
 
-    try
+    if (auto isTestDataExist = UserTable::isUsernameExist(rootName);
+        isTestDataExist != nullptr && (*isTestDataExist))
     {
-        futureContact.get();
-        futureUser.get();
+        LOG_INFO << "Test data already exists.";
+
+        auto dbClient = drogon::app().getDbClient();
+
+        auto futureResult = dbClient->execSqlAsyncFuture(
+            "SELECT id, username FROM \"User\"");
+
+        try
+        {
+            auto result = futureResult.get();
+
+            Json::Value userIds(Json::arrayValue);
+
+            for (const auto& row : result)
+            {
+                const std::string id = row["id"].as<std::string>();
+                LOG_DEBUG << row["username"].as<std::string>() << ": "
+                          << id << ", token: " << genJwtToken(id);
+            }
+        }
+        catch (const drogon::orm::DrogonDbException &e)
+        {
+            LOG_ERROR << "Database error: " << e.base().what();
+        }
+
+        return;
     }
-    catch (const orm::DrogonDbException &e)
+    else
     {
-        LOG_ERROR << "Debug setup: Database error: " << e.base().what();
+        LOG_WARN << "TEST DATA IS CORRUPTED. CLEARING TABLES.";
+        auto dbClient = app().getDbClient();
+        auto futureContact = dbClient->execSqlAsyncFuture("TRUNCATE TABLE \"Contact\" CASCADE;");
+        auto futureUser= dbClient->execSqlAsyncFuture("TRUNCATE TABLE \"User\" CASCADE;");
+
+        try
+        {
+            futureContact.get();
+            futureUser.get();
+        }
+        catch (const orm::DrogonDbException &e)
+        {
+            LOG_ERROR << "Debug setup: Database error: " << e.base().what();
+            return;
+        }
+    }
+
+    Json::Value testData;
+    testData["name"] = rootName;
+    testData["username"]  = rootName;
+    testData["password"] = "test_password";
+
+    UserTable rootUser(std::make_shared<Json::Value>(testData));
+    auto rootUserData = rootUser.addNewUser();
+
+    if (rootUserData == nullptr)
+    {
+        LOG_ERROR << "Failed to add root user";
         return;
     }
 
-    std::string testName = "test_name";
-    std::string testUsername = "test_username";
+    const std::string testName = "test_name";
+    const std::string testUsername = "test_username";
 
-    Json::Value testData;
     testData["name"] = testName;
     testData["username"]  = testUsername;
-    testData["password"] = "test_password";
 
     UserTable primaryUser(std::make_shared<Json::Value>(testData));
     auto primaryUserData = primaryUser.addNewUser();
@@ -116,10 +170,10 @@ static void addTestData()
         return;
     }
 
-    std::string primaryUserId = (*primaryUserData)["id"].asString();
+    const std::string primaryUserId = (*primaryUserData)["id"].asString();
 
-    LOG_INFO << "test_username id: " << primaryUserId
-             << " token: " << genJwtToken(primaryUserId);
+    LOG_DEBUG << "test_username id: " << primaryUserId
+              << " token: " << genJwtToken(primaryUserId);
 
 
     for (int i = 1; i <= 15; i++)
@@ -136,10 +190,10 @@ static void addTestData()
             continue;
         }
 
-        std::string testUserId = (*testUserData)["id"].asString();
+        const std::string testUserId = (*testUserData)["id"].asString();
 
-        LOG_INFO << testData["name"].asString() << " id: " << testUserId
-                 << " token: " << genJwtToken(testUserId);
+        LOG_DEBUG << testData["name"].asString() << " id: " << testUserId
+                  << " token: " << genJwtToken(testUserId);
 
         if (i <= 12)
         {
@@ -148,8 +202,8 @@ static void addTestData()
     }
 
     auto fakeUserId = utils::getUuid(false);
-    LOG_INFO << "fake_user id: " << fakeUserId << " token: " << genJwtToken(fakeUserId);
+    LOG_DEBUG << "fake_user id: " << fakeUserId << " token: " << genJwtToken(fakeUserId);
 
     LOG_INFO << "Adding test data completed.";
 }
-#endif
+#endif // !NDEBUG
